auth-proxy/pyproject.toml
[project]
name = "auth-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.0",
    "python-dotenv>=1.2.1",
]
